Initial tests with unrolling rabinkarp_update() as a function.
This adds rabinkarp.c copying rollsum.c's unrolled loop idea.
